fixes issues with network switch pop up
